requirements.txt
anyio==3.7.0
APScheduler==3.6.3
blinker==1.6.2
cachetools==4.2.2
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
Flask==2.3.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.33
protobuf==3.20.3
python-telegram-bot==13.13
pytz==2023.3
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
Werkzeug==2.3.4
